requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
sqlparse==0.4.2
django-cors-headers==3.11.0
django-environ==0.4.5
djangorestframework==3.12.2
Pillow==8.3.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pytz==2020.5
six==1.15.0
toml==0.10.2
